+++ source/common/http/async_client_impl.h
@@ -278,7 +278,7 @@ private:
   Tracing::NullSpan active_span_;
   const Tracing::Config& tracing_config_;
   const LocalReply::LocalReply& local_reply_;
-  const std::unique_ptr<const Router::RetryPolicy> retry_policy_;
+  Router::RetryPolicyConstSharedPtr retry_policy_;
   std::shared_ptr<NullRouteImpl> route_;
   uint32_t high_watermark_calls_{};
   bool local_closed_{};
