+++ source/common/router/BUILD
@@ -43,7 +43,7 @@ envoy_cc_library(
         ":matcher_visitor_lib",
         ":metadatamatchcriteria_lib",
         ":per_filter_config_lib",
-        ":reset_header_parser_lib",
+        ":retry_policy_lib",
         ":retry_state_lib",
         ":router_ratelimit_lib",
         ":tls_context_match_criteria_lib",
@@ -539,5 +539,21 @@ envoy_cc_library(
     deps = [
         ":delegating_route_lib",
         "//envoy/router:cluster_specifier_plugin_interface",
+    ],
+)
+
+envoy_cc_library(
+    name = "retry_policy_lib",
+    srcs = ["retry_policy_impl.cc"],
+    hdrs = ["retry_policy_impl.h"],
+    deps = [
+        ":reset_header_parser_lib",
+        ":retry_state_lib",
+        "//envoy/router:router_interface",
+        "//envoy/server:factory_context_interface",
+        "//source/common/config:utility_lib",
+        "//source/common/upstream:retry_factory_lib",
+        "@com_google_absl//absl/types:optional",
+        "@envoy_api//envoy/config/route/v3:pkg_cc_proto",
     ],
 )
